+++ feature/profile-edit/impl/src/test/java/ru/yeahub/profile_edit/impl/presentation/profile_edit_screen_mapper_tests/ProfileEditScreenMapperErrorThrowableMappingTest.kt
@@ -1,4 +1,4 @@
-package ru.yeahub.profile_edit.impl.presentation.profile_edit_mapper_tests
+package ru.yeahub.profile_edit.impl.presentation.profile_edit_screen_mapper_tests
 
 import okhttp3.ResponseBody.Companion.toResponseBody
 import org.junit.jupiter.api.Assertions.assertEquals
