+++ feature/profile-edit/impl/src/test/java/ru/yeahub/profile_edit/impl/presentation/profile_edit_screen_mapper_tests/ProfileEditScreenMapperStateTransitionTest.kt
@@ -1,4 +1,4 @@
-package ru.yeahub.profile_edit.impl.presentation.profile_edit_mapper_tests
+package ru.yeahub.profile_edit.impl.presentation.profile_edit_screen_mapper_tests
 
 import kotlinx.collections.immutable.persistentListOf
 import kotlinx.collections.immutable.persistentMapOf
